Exceptions and errors
How to handle observable errors and error thrown otherwise.
How to use catch() to handle exceptions.
How to use onErrorResumeNext operator instead of merge to continue with next observable item on error.
